backend/requirements/requirements_production.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
defusedxml==0.7.1
Django==3.2.3
django-filter==21.1
djangorestframework==3.12.4
djoser==2.3.1
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
hashids==1.3.1
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python3-openid==3.2.0
python_dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
gunicorn==20.1.0